Followup of https://github.com/iotappstory/ESP-Library/pull/30
Re-implemented changes which we removed out of #30 in order to not introduce breaking changes. Since we are now making a new major release this is the right time to introduce such changes:

I've created a general `IOTAppStory::loop()` function for the library which will do 2 things:
- call `buttonLoop()` function every time (so this is not necessary in the main sketch anymore)
- perform call home functionality if it is enabled, and at the correct interval that was configured.

The `IOTAppStory::loop()` function should be called in the main sketch's `loop()`.

I've also modified all examples to represent these changes.
README.md has been updated as well