urlencode quotes in value Strings before sending
- replace " with %22
- decoding is done browser side

Saving Strings from the config pages with quotes to EEPROM is not a problem nor using them in your Sketch.
However when sending them with other values in a json String to the browser(config pages) these extra quotes break the json format.
This is why we now urlencode the quotes before sending. The config page js scripts take care of decoding.