fix: ESP8266 exit config crash
When entering or leaving config mode your device will send this and the local ip address in a log message to the IAS server.

This log process uses BearSSL and would run out of heap when leaving config mode. (the second time it would run)

This is fixed by adding"scope" { } which releases the heap used by the first log and the config server.