Add libraryUpgrader example
This sketch is an example of how you can upgrade from the previous library(2.0.X) to the new 2.1.0 library EEPROM layout. If successful all the previous settings like Wifi, IAS device registration and added fields get "rewritten" to the new EEPROM layout. 

You can use this sketch as an intermediate update or add this code to your next firmware update. It's wise to test this locally first! Make sure to include the upgrade helper file in your sketch folder. And do NOT use presets!