Only connect to Wifi if there is NO connection (#169)
I have notice an issue when calling IOTAppStory::begin() when wifi is already configured on an ESP32. The connection is disconnected and IAS attempts to connect using its own internal routines. This can be annoying and interfere with the authors intent, By only calling when there is no connection, this can be avoided, without and greater impact to the library functionality.

##TLDR
Only connect if not *already* connected from some action before the IOTAppStory::begin() call. This helps avoids wifi managers interfering with each other - it's a quick hack but *seems* to work just fine.