Update IOTAppStory.h
Added 
- public void preSetConnWifiOnBoot(bool automaticWifiConnectOnBoot);
- private bool _automaticWifiConnectOnBoot
This lets programmers decide whether or not to setup and initiate wifi on boot. (in IAS.begin())

Changed
- private bool _connected to public bool WiFiConnected
This lets programmers check if the device is connected to an WiFi AP. 
Or if they choose to use their own WiFi setup solution, tell this library if the device is connected.

Added a precompiler check for ARDUINO_SAMD_VARIANT_COMPLIANCE whether or not to compile the  servSaveAppInfo() function.

Changed the configmode station server js address. For backward compatibility the old address is still accepted by our servers!