Update IOTAppStory.cpp
Added 
- public void preSetConnWifiOnBoot(bool automaticWifiConnectOnBoot);
- private bool _automaticWifiConnectOnBoot
This lets programmers decide whether or not to setup and initiate wifi on boot. (in IAS.begin())

Changed
- private bool _connected to public bool WiFiConnected
This lets programmers check if the device is connected to an WiFi AP. 
Or if they choose to use their own WiFi setup solution, tell this library if the device is connected.

Fixed bad code
- preSetWifi

NTP
- added ntpSync(int retries)
- changed setClock();

Updated the code to the new var & function names named above.