Multiple updates, additions & improvements (#171)
* Update config.h

Added defines for the default port nrs:
- OTA_HOST_HTTP_PORT 80
- OTA_HOST_HTTPS_PORT 443

* Update config.h

Changed default values for OTA_UPD_FILE & OTA_LOG_FILE
For backward compatibility the old addresses are still supported!

* Update CallServer.cpp

Changed hardcode port nrs to defines which are defined in the config.h file.

* Update CallServer.cpp

Changed hardcode port nrs to defines which are defined in the config.h file.

* Update WiFiConnector.h

Introduced a new private var _nrOfCfgFound
And made sure that Arduino compiler only compiles when ESP8266 & ESP32 are defined to prevent errors for boards like the Ard. Wifi MKR 1010

* Update WiFiConnector.cpp

Make sure that it does not try to connect when 0 credentials are found using the private var _nrOfCfgFound

And made sure that Arduino compiler only compiles this file when ESP8266 & ESP32 are defined to prevent errors for boards like the Ard. Wifi MKR 1010

* Update IOTAppStory.h

Added 
- public void preSetConnWifiOnBoot(bool automaticWifiConnectOnBoot);
- private bool _automaticWifiConnectOnBoot
This lets programmers decide whether or not to setup and initiate wifi on boot. (in IAS.begin())

Changed
- private bool _connected to public bool WiFiConnected
This lets programmers check if the device is connected to an WiFi AP. 
Or if they choose to use their own WiFi setup solution, tell this library if the device is connected.

Added a precompiler check for ARDUINO_SAMD_VARIANT_COMPLIANCE whether or not to compile the  servSaveAppInfo() function.

Changed the configmode station server js address. For backward compatibility the old address is still accepted by our servers!

* Update config.h

Added defines for the configTime function

* Update IOTAppStory.h

Added default value for
ntpSync(int retries = SNTP_CONN_MAX_RETRIES);

Removed redundant comments

* Update IOTAppStory.cpp

Added 
- public void preSetConnWifiOnBoot(bool automaticWifiConnectOnBoot);
- private bool _automaticWifiConnectOnBoot
This lets programmers decide whether or not to setup and initiate wifi on boot. (in IAS.begin())

Changed
- private bool _connected to public bool WiFiConnected
This lets programmers check if the device is connected to an WiFi AP. 
Or if they choose to use their own WiFi setup solution, tell this library if the device is connected.

Fixed bad code
- preSetWifi

NTP
- added ntpSync(int retries)
- changed setClock();

Updated the code to the new var & function names named above.

* Update serialFeedback_EN.h

* Update serialFeedback_EN.h

Added timeout message for configmode

* Update config.h

Added CFG_TIMEOUT define to tell Config mode to Exit after X time of inactivity.

* Update ConfigServer.h

Added unsigned int _lastAction 
last action taken by human to determine when to timeout

* Update ConfigServer.cpp

Added
- Config mode timeout. Exit after X time of inactivity. Set CFG_TIMEOUT in the config.h file
- Missing writeConfig() for storing firmware name & version after local(browser) upload.

Updated
- private bool _connected to public bool WiFiConnected

* Update IOTAppStory.cpp

Allow configmode announcement when stored in SPIFFS

* Update CallServer.h

Replaced PROGMEM stored certificate with the latest.

* Update CallServer.h

Replaced PROGMEM stored certificate with the latest.

* Update ConfigServer.cpp

Added missing CFG_TIMEOUT define instead of hardcoded nr

* Update and rename examples/IASLoader/IASLoader.ino to examples/Loaders/IASLoader/IASLoader.ino

On first boot copy hardcoded root certificate to a file in SPIFFS /cert/iasRootCa.cer
This saves the user time as users often forget to do this in config mode the first time.

* Updated certificate fingerprint to the latest version

* Update and rename examples/Other Examples/addDeviceToProject/addDeviceToProject.ino to examples/Loaders/BatchLoader/BatchLoader.ino

- Renamed to Batchloader
- Updated comments
- Updated root certifcate & fingerprint
- Put client connect into a loop untill connected
- Fixed typo

* Fixed typo

* Added ntpSync keyword

* increase version nr

* increase version nr

* Increase version nr

* Added missing define

* Sync changes with English file

* Update ConfigServer.cpp

Added temp workaround between setting softAP() and softAPConfig() to prevent crashes on the ESP32 when connecting.